ClanWars Admin section done

+++ app/Http/Traits/Replay.php
@@ -18,36 +18,56 @@ trait Replay
         $this->replay = new stdClass();
         $this->replay->Errors = [];
 
-        $executeScrep = new Process(['screp.exe', $filePath]);
-        $executeScrep->run();
+        switch (config('server.type')) {
+            case 'LINUX':
+                $executeScrep = new Process(['./screp', $filePath]);
+                break;
+            
+            case 'WINDOWS':
+                $executeScrep = new Process(['screp.exe', $filePath]);
+                break;
 
-        if (!$executeScrep->isSuccessful()) {
-
-            $this->removeFile($filePath) ? null : abort(403, 'Contact administrator, unable to remove file!');
-
-            if (env('APP_ENV') == 'local') {
-                throw new ProcessFailedException($executeScrep);
-                exit(1);
-            }
-
-            if ($executeScrep->getOutput() == "Failed to parse replay: Decoder.Section() error: not a replay file\n")
-            {
-                array_push($this->replay->Errors, 'Not a replay!');
-                return $this->replay;
-            } else
-            {
-                array_push($this->replay->Errors, 'Invalid file.');
-                return $this->replay;
-            }
+            case 'HOSTING':
+                # code
+                break;
+            
+            default:
+                $executeScrep = new Process(['./screp', $filePath]);
+                break;
         }
-        
-        $this->replay->Data = json_decode($executeScrep->getOutput());
 
-        if ($this->checkReplayEngine($this->replay->Data) == false)
+        if((config('server.type') == 'LINUX') OR (config('server.type') == 'WINDOWS'))
         {
-            array_push($this->replay->Errors, 'Invalid engine.');
-            $this->removeFile($filePath) ? null : abort(403, 'Contact administrator, unable to remove file!');
-            return $this->replay;
+            $executeScrep->run();
+
+            if (!$executeScrep->isSuccessful()) {
+    
+                $this->removeFile($filePath) ? null : abort(403, 'Contact administrator, unable to remove file!');
+    
+                if (env('APP_ENV') == 'local') {
+                    throw new ProcessFailedException($executeScrep);
+                    exit(1);
+                }
+    
+                if ($executeScrep->getOutput() == "Failed to parse replay: Decoder.Section() error: not a replay file\n")
+                {
+                    array_push($this->replay->Errors, 'Not a replay!');
+                    return $this->replay;
+                } else
+                {
+                    array_push($this->replay->Errors, 'Invalid file.');
+                    return $this->replay;
+                }
+            }
+            
+            $this->replay->Data = json_decode($executeScrep->getOutput());
+    
+            if ($this->checkReplayEngine($this->replay->Data) == false)
+            {
+                array_push($this->replay->Errors, 'Invalid engine.');
+                $this->removeFile($filePath) ? null : abort(403, 'Contact administrator, unable to remove file!');
+                return $this->replay;
+            }
         }
 
         return $this->replay;
